deps: Update fmtlib to 11.0.2 (#35869)
Commit Message:
fmtlib no longer supports implicit conversion from enum to int and as
such all call sites needed to be modified to explicitly convert enum
values to int.

---------

+++ source/extensions/load_balancing_policies/common/thread_aware_lb_impl.h
@@ -38,7 +38,8 @@ public:
           host->metadata().get(), Config::MetadataFilters::get().ENVOY_LB,
           Config::MetadataEnvoyLbKeys::get().HASH_KEY);
       if (val.kind_case() != val.kStringValue && val.kind_case() != val.KIND_NOT_SET) {
-        FINE_GRAIN_LOG(debug, "hash_key must be string type, got: {}", val.kind_case());
+        FINE_GRAIN_LOG(debug, "hash_key must be string type, got: {}",
+                       static_cast<int>(val.kind_case()));
       }
       absl::string_view hash_key = val.string_value();
       if (hash_key.empty()) {
